requirements.txt
apache-airflow
pandas
beautifulsoup4